Finish

+++ backend/src/controllers/notes.ts
@@ -1,34 +1,48 @@
 import { RequestHandler } from "express";
-import NoteModel from "../models/notes"
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import NoteModel from "../models/note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec(); // using exec to turn to real promise
+        assertIsDefined(authenticatedUserId);
+
+        const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
         res.status(200).json(notes);
     } catch (error) {
-        next(error)
+        next(error);
     }
-}
+};
 
 export const getNote: RequestHandler = async (req, res, next) => {
-    const noteId = req.params.noteId
+    const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
+
         const note = await NoteModel.findById(noteId).exec();
 
         if (!note) {
-            throw createHttpError(404, "Note not found")
+            throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         res.status(200).json(note);
     } catch (error) {
-        next(error)
+        next(error);
     }
-}
+};
 
 interface CreateNoteBody {
     title?: string,
@@ -38,23 +52,26 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
-    
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!title) {
-            throw createHttpError(400, "Note must have title")
+            throw createHttpError(400, "Note must have a title");
         }
 
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
-        })
+        });
 
-        res.status(201).json(newNote)
-        
+        res.status(201).json(newNote);
     } catch (error) {
-        next(error)
+        next(error);
     }
-}
+};
 
 interface UpdateNoteParams {
     noteId: string,
@@ -66,24 +83,30 @@ interface UpdateNoteBody {
 }
 
 export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
-    const noteId = req.params.noteId
+    const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
-    
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
         }
 
         if (!newTitle) {
-            throw createHttpError(400, "Note must have title")
+            throw createHttpError(400, "Note must have a title");
         }
 
         const note = await NoteModel.findById(noteId).exec();
 
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
         }
 
         note.title = newTitle;
@@ -91,17 +114,18 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         const updatedNote = await note.save();
 
-        res.status(200).json(updatedNote)
-        
+        res.status(200).json(updatedNote);
     } catch (error) {
-        next(error)
+        next(error);
     }
-}
+};
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
-    const noteId = req.params.noteId
-    
+    const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
     try {
+        assertIsDefined(authenticatedUserId);
 
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id");
@@ -113,10 +137,14 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
             throw createHttpError(404, "Note not found");
         }
 
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note");
+        }
+
         await note.deleteOne();
 
-        res.sendStatus(204);        
+        res.sendStatus(204);
     } catch (error) {
-        next(error)
+        next(error);
     }
-}
+};
